Refactor structure of the backtest directory
