+++ trader/backtest/analysis/__init__.py
@@ -1,0 +1,1 @@
+from trader.backtest.analysis.analyzer import StockBacktestAnalyzer
