+++ trader/backtest/analysis/base.py
@@ -4,20 +4,6 @@ from typing import Optional
 from trader.models import StockAccount
 from trader.strategies.stock import BaseStockStrategy
 
-"""
-base.py
-
-Defines abstract base classes for performance-related modules.
-
-This file provides reusable interfaces for backtest performance analysis and report generation.
-Typical use cases include customizing analyzers or reports for different financial instruments
-(e.g. stocks, futures, options).
-
-Classes:
-- BaseAnalyzer: Interface for computing backtest metrics
-- BaseReport: Interface for generating performance reports
-"""
-
 
 class BaseBacktestAnalyzer(ABC):
     """Backtest Performance Analyzer Framework (Base Template)"""
@@ -89,37 +75,4 @@ class BaseBacktestAnalyzer(ABC):
     @abstractmethod
     def compute_num_losing_trades(self) -> int:
         """計算虧損筆數（可用於 win rate）"""
-        pass
-
-
-class BaseBacktestReporter(ABC):
-    """Backtest Performance Reporter Framework (Base Template)"""
-
-    def __init__(self, strategy: BaseStockStrategy):
-        self.strategy: BaseStockStrategy = strategy  # Backtest strategy
-        self.account: StockAccount = self.strategy.account  # Account
-
-    @abstractmethod
-    def plot_equity_curve(self) -> None:
-        """計算並繪製權益曲線（淨資產隨時間變化）"""
-        pass
-
-    @abstractmethod
-    def plot_equity_and_benchmark_curve(self) -> None:
-        """計算並繪製權益 & benchmark 曲線圖"""
-        pass
-
-    @abstractmethod
-    def plot_mdd(self) -> None:
-        """計算並繪製 Max Drawdown"""
-        pass
-
-    @abstractmethod
-    def set_figure_config(self) -> None:
-        """設置繪圖配置"""
-        pass
-
-    @abstractmethod
-    def save_figure(self) -> None:
-        """儲存回測報告"""
         pass
