--- trader/backtest/performance/__init__.py
@@ -1,15 +1,0 @@
-from .analyzer import StockBacktestAnalyzer
-from .base import BaseBacktestAnalyzer
-
-"""
-performance package
-
-This package contains modules for analyzing and reporting backtest performance.
-It includes reusable components for computing key metrics (e.g., ROI, MDD),
-generating reports, and supporting asset-specific customizations.
-
-Modules:
-- analyzer: Performance metric calculators (e.g., StockAnalyzer)
-- report: Report generator for visualizing and exporting results
-- results/: Folder for storing output files (e.g., charts, Excel summaries)
-"""
