+++ trader/backtest/report/__init__.py
@@ -1,0 +1,1 @@
+from trader.backtest.report.reporter import StockBacktestReporter
